requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

atproto==0.0.56
ffmpeg-python==0.2.0
pillow==11.0.0
praw==7.7.1
prawcore==2.4.0
python-dotenv==1.0.1
requests==2.32.3
